Allow dispatcher to shutdown after a number of jobs (executions) performed
This was suggested by Donal (thanks!) as a form of memory management, but also
because with Sidekiq workers checked for orphan claimed jobs when they started, and
if the workers that held them had died not that long ago so that their heartbeat
hand't expired, the jobs wouldn't be released. In Solid Queue this doesn't happen
because the supervisor checks for claimed jobs to release periodically, but we
might need this for memory management purposes or for other reasons. It's off
by default.