Make Configuration to return the instantiated runners
In advance of actually getting configured runners simply from it, and
doing the real instantiation in that new configured runner class, or in
the supervisor. The goal is to not reuse the same object when replacing a
dead child.

Also: test sending multiple signals, remove unnecessary call to Process.setpgrp
(we don't need to set different groups there, at least not for now) and tidy up
some more Supervisor code.