Run solid_queue with SQLite
Update the tests to run solid_queue on SQLite.

They can be triggered by:
```
TARGET_DB=sqlite rails test
```

The adapter is prone to raising `SQLite3::BusyException`s when
concurrent transactions occur. Preventing this requires a couple of
patches to the adapter.

1. Implement retry with backoff - add a `retries` config setting
to the adapter and sleep progressively longer for each retry. This
setting is currently in the Rails main branch, but that implementation
has no backoff. That doesn't work well in our case.
1. Always create immediate transactions - SQLite by default creates
deferred transactions, which don't take a write lock. Then later if
there is a write it tries to upgrade the lock. This won't work if the
transaction has a stale read so retrying the write by itself is not
possible. Starting with an immediate transaction moves the write lock to
that point and ensures that we only get blocked on a retryable
transaction.