Delete ready_executions only when they've been claimed
If we try to claim a set of job IDs, creating claimed executions for them,
but for some reason, we fail to do that for some of them, in a way that the
failure is ignored (eg. on a duplicate key conflict that gets ignored with
Rails's `insert_all`), we'd be deleting the ready execution for these, even
if they weren't claimed at all, leaving the job in limbo. This change simplifies
a bit the claiming code and ensures only those jobs that have been claimed are
indeed deleted.