Implement basic concurrency control on enqueuing using semaphores
At the time of enqueuing a job and creating a ready execution for it, we check
if the job has concurrency control settings. If it has, we use a semaphore to
determine whether the job can go to the ready_executions table or whether it needs
to wait in the new blocked_executions table. The semaphore is created for the job
concurrency key and has just a value that gets decremented avoiding race conditions,
only if it's still > 0.

This needs to be paired with jobs releasing/signaling the semaphore when they finish,
and the worker doing a pass over blocked executions in addition to polling.