Dispatch blocked jobs after job perform finishes outside that transaction
Otherwise, if a job finishes just fine or fails, and something goes wrong when
dispatching blocked jobs, we'll fail to mark the job as finished or failed just
because we couldn't dispatch blocked jobs, but not because anything went wrong with the
job itself.

Also: test sequential jobs when some of them fail. We must continue processing jobs
normally.