Correct transactions to release blocked executions
When a job finishes or fails, we need to:
- Increment the semaphore if any, in its own transaction. This needs to
happen regardless of whether other blocked executions are unblocked by this
semaphore change.
- Try to unblock the next blocked execution. For this, we need to acquire the
semaphore again, as it's possible another job part of the same concurrency group
is just being enqueued at the same time. Then, we need to move the blocked
execution to "ready", and then delete the blocked execution. This all needs to
happen in the same transaction, without going throuhg the job. The previous
implementation could very well update the existing blocked execution and then
delete it, leaving the job in limbo.