Add support for concurrency controls across jobs (aka. concurrency "groups")
That's it, if we wanted for example, to limit concurrency for a group of jobs together
because they act on the same kind of record, and thus, we can't have the job class be part
of the concurrency key, as that would keep them all separated. With this change, we have two
parts to build the concurrency key:
- The concurrency group, which defaults to the job class name
- And the concurrency key itself, that defaults to the first argument

This also fixes the previous calculation of the key, that was working by chance because it was
being executed in the context of the included module, not the job class.