Enqueue jobs using a different DB connection from the one checked-out
If there's a connection checked-out, and if we're in a transaction. The reason to do
this is avoiding enqueing the job in the same transaction as whatever the caller is
doing, so we don't silently make the caller rely on transactional integrity when
enqueuing jobs. Later, we'll add a way to override this so we can rely explicitly on
transactional integrity where it makes sense.