Only try to release next unblocked job if the job actually completed
Either successfully or failing, but it needs to have completed. Otherwise,
we might be still claimed but signal the semaphore regardless, so it'd be
lying about how many jobs are in progress. A good example where this might
happen is when the worker is sent a QUIT signal to exit right away, and the
job gets an exit! midway. As the worker would try to release claimed executions
after the shutdown, the claimed execution that holds the semaphore could be
potentially blocked because the semaphore is held at least by itself. Then,
depending on the order of the thread pool shutting down and the worker being
deregistered, we could end up with a job trying to unblock itself and a stuck
semaphore.