Start discard from execution, not job
To avoid a race condition when we check whether we have a ready execution and the job
gets claimed right before, and after we've already checked if it was claimed. Rely
always on executions to delete a job that's not finished. When it's finished, it can
be cleared in bulk, without having to worry about the callbacks.