Stop retrying semaphore decrement when getting a deadlock
This was added in d72fe4607f8fc5aacc07c6314fb5aa0af200d6f0 and initially,
it seemed like a good idea, but it ended up causing a different error
as it messed with Active Record's transactions. The error would be
```
ActiveRecord::StatementInvalid (Mysql2::Error: SAVEPOINT active_record_1 does not exist)
```
And this made us realise it might be more complicated, because we might be
in a transaction that wraps the Solid Queue transaction.

Since this is such an edge case that happens only with certain job loads
and perhaps only with MySQL, InnoDB and READ COMMITTED, let's stop trying to
be too clever about it, and instead, simplify this a bit by avoiding the
decrement at all when the semaphore's limit is 1 and it was just created.