Support passing kwargs as last element of the arguments array
If we don't explicitly add a ruby2_keywords flag, Active Job will any hash
included in the arguments array with keys as `_aj_symbol_keys`, and when
deserialized, it'd be treated always as a hash argument instead of keyword
arguments. Depending on the job, this might work fine, but if the job uses
keyword arguments, trying to execute the job with deserialized arguments
will fail. However, the opposite is not true: if the job accepts a hash
argument and we pass a hash with the ruby2_keywords flag, it'll work just
fine as Active Job will serialize that with keys as `_aj_ruby2_keywords`,
so we take advantage of that to simplify the task definition and not
having to distinguish between args and kwargs.