Refactor a bit the Poller vs. Runnable modules
Make the loop be part of Poller. Allow to have other other
Runnable processes that don't need an infinite loop.

I'm still not super happy with these concerns. This needs more
work that will come when I properly implement async mode. Right now
this is all interleaved in the modules and it shouldn't be.