Support multiple databases for install generator
Currently, there is no way to specify that the migration files be installed into a separate database from the default. This is problematic especially for SQLite apps, since the recommended approach is to use a separate SQLite database for the queue. 

This change supports the existing behavior as well as specifying the `--database` option or the `--db` option.