Release any orphaned claimed executions as part of the process prune
Orphaned executions shouldn't happen as part of regular operation but
some people have reported them. The only way I can think of this
happening could be that a worker gets deregistered while still working
normally, so that it happens to be claiming executions while the process
record is deleted. This means that the callback to release claimed
executions might not see yet the executions that are being claimed at
the same time, and as such, it won't release any. Or, in other words,
those executions would get claimed by a process that's getting deleted
and that once they're committed to the DB, it no longer exists.

To recover from this scenario, as well as checking for processes with an
expired heartbeat, the supervisor will also check for orphaned
executions and release them.