quick grammar fixup in README.md
Just a few small tweaks... feel free to ask me to explain these changes if the reasons are not clear.  I'm happy to go into as much or as little depth as you'd like, either in comments on the PR or in a Zoom call or whatever else works best for you.