Fix bug when enqueuing the job for a recurring task fails
As `perform_later` in this case returns `false`, which crashes when we
try to access `job_id` on it. `perform_later` returns the job instance
with `successfully_enqueued` set to true when it goes well, but `false`
when it fails. You can provide a block and the job instance with
`successfully_enqueued` set to `false` and an error set in
`enqueue_error` is yielded to that block, but it's not returned.

At first I tried to make this a bit more sophisticated, passing a block
to get the job and populate the instrumentation payload with it, but the
code was getting quite cumbersome and the error is instrumented by
Active Job anyway, so in the end I just simplified this and opted for
handling this case to avoid crashing.