Raise a custom exception instead of ActiveJob::EnqueueError on enqueue
The reason is that we can't easily control what happens when something
goes wrong on enqueuing if the jobs are enqueued by Rails or other gems
(eg. Turbo::Streams::BroadcastJob, ActiveStorage::AnalyzeJob...),
because the only way is to check what the return value of the call to
`perform_later` is, or to pass a block to that method to fetch the error
that gets set in `enqueue_error`. In this way at least the error will
bubble up and you can rescue or let it be raised.