Truncate error backtrace when saving failed executions
Otherwise it can get really big for SystemStackError errors, and not fit
in the DB text column.

Assuming an average of 150 characters per backtrace line, limting to 400
would be around 60K characters plus the error name and error message,
that are usually small, and the JSON attributes and other JSON
serialization characters.

This is the first approach to this. I'll change this to more
sophisticated truncation that checks the column size and calculates how
much it needs to truncate instead of using a fixed number.