Handle concurrent inserts in PostgreSQL when enqueuing recurring tasks
Very similar to inserting recurring tasks on boot.

PostgreSQL fails and aborts the current transaction when it hits
a duplicate key conflict during two concurrent INSERTs for the same
value of an unique index. We need to use `insert` instead of `create`
here and indicate unique_by to ignore duplicate rows by this value
when inserting