Create always new Worker and Dispatcher instances before starting
We were reusing the instances of Worker and Dispatcher from the initial
configuration all the time, which could bring some problems with stopped
pools. Now that we need a name to be generated and be unique per process
instance, we really need to instantiate new processes every time they're
started.