Support recurring jobs that defer enqueuing to after commit always
These need to be enqueued within the transaction that creates the
recurring execution to avoid duplicate crons. However, if the job is
set to `enqueue_after_transaction_commit = :always` or :default when we
change the default, it won't work because we'll try to create the
recurring execution without having actually enqueued the job. With this
change, we bypass Active Job's enqueuing and enqueue directly with Solid
Queue, running enqueue callbacks.