Don't try to start a supervisor on an empty configuration
It leads to weird errors, especially when forking, such as
`Errno::ECHILD` when trying to check on the children status without
having forked any children.

I want to improve this with a generic configuration validation step,
that would check this and other things, and abort/raise if there are any
problems, but this will do for now.