Fix issue when pruning a supervisor and its supervisees via callbacks
In that case we weren't correctly failing executions as a consequence of
the prune because the processes would get deleted via a callback without
going through the regular prune. Leave this to the prune itself instead
of relying on the callbacks and only deregister supervised processes
when not pruning. This also saves a query because we check whether we
have a supervisor or not before trying to find supervisees.