Fix enqueuing recurring task with job outside Solid Queue
Calling `set` on the job class returns a `ConfiguredJob`, and calling
`perform_later` on a `ConfiguredJob` doesn't yield the job to a block
passed as parameter, so we can't check if the enqueuing was successful
or not. To work around this, the only way is to instantiate the job
ourselves with the arguments, then call `enqueue` on that, and rely on
the result, which would be what `job_class.perform_later` would yield
to the block.