Support Active Job Continuations
[Active Job Continuations](https://github.com/rails/rails/pull/55127)
will be added in Rails 8.1. Queue adapters need to implement the
`stopping?` method to support them.

Implement the method and set it to true when a worker is stopping.

Also add Rails versions so we can test against Rails main. To prevent an
explosion of combinations, I've switched to only testing against minor
versions of Ruby.