Set `successfully_enqueued` in `Job.enqueue`
Even though this will be overwritten by Active Job when called through
`perform_later`.

This is necessary for recurring tasks, where we run the `enqueue`
callbacks manually and call `Job.enqueue`. We rely on `successfully_enqueued?`
to know whether we need to record the recurring execution. We were setting
that no matter what after calling `Job.enqueue` because we were relying on
that to raise (which would be the case for Active Record errors).

In the case of discarding it because of concurrency controls, the job is simply
not persisted but no error is raised, which means we were trying to
record a recurring execution without a persisted job, which would raise.

Fixes #598