Make concurrency tests less flaky
Don't try to enforce sequential order for jobs, given that the order is
not even guaranteed by Solid Queue. Rename the test job to reflect that.
Fix also another test where we waited for a job to finish but not for it
to release the semaphore, so it happened most of the times but sometimes
it didn't.